@@ -30,23 +30,26 @@
         }
     ],
     "require": {
-        "php": "^7.2",
+        "php": "^8.0",
         "ext-json": "*",
         "ext-intl": "*",
-        "dompdf/dompdf": "^0.8.0",
-        "illuminate/database": "^6.0|^7.0|^8.0",
-        "illuminate/support": "^6.0|^7.0|^8.0",
+        "dompdf/dompdf": "^1.0",
+        "illuminate/database": "^8.0",
+        "illuminate/support": "^8.0",
         "mollie/laravel-mollie": "^2.9",
+        "mollie/mollie-api-php": "^2.27",
         "moneyphp/money": "^3.2",
-        "nesbot/carbon": "^2.0"
+        "nesbot/carbon": "^2.31"
     },
     "require-dev": {
+        "guzzle/http": "^3.9",
+        "guzzlehttp/guzzle": "^7.0",
         "laravel/legacy-factories": "^1.0",
-        "mockery/mockery": "^1.0",
-        "orchestra/testbench": "^4.0|^5.0|^6.0",
-        "phpunit/phpunit": "^8.3|^9.0",
-        "vlucas/phpdotenv": "^3.3|^4.0|^5.0",
-        "friendsofphp/php-cs-fixer": "^v2.17"
+        "mockery/mockery": "^1.4",
+        "orchestra/testbench": "^6.0",
+        "phpunit/phpunit": "^9.3",
+        "vlucas/phpdotenv": "^5.2",
+        "friendsofphp/php-cs-fixer": "^3.0"
     },
     "autoload": {
         "files": [
