@@ -7,11 +7,12 @@ use Laravel\Cashier\FirstPayment\Actions\AddGenericOrderItem;
 use Laravel\Cashier\FirstPayment\FirstPaymentBuilder;
 use Laravel\Cashier\Mollie\Contracts\CreateMollieCustomer;
 use Laravel\Cashier\Mollie\Contracts\CreateMolliePayment;
+use Laravel\Cashier\Payment;
 use Laravel\Cashier\Tests\BaseTestCase;
 use Laravel\Cashier\Tests\Fixtures\User;
 use Mollie\Api\MollieApiClient;
 use Mollie\Api\Resources\Customer;
-use Mollie\Api\Resources\Payment;
+use Mollie\Api\Resources\Payment as MolliePayment;
 use Mollie\Api\Types\SequenceType;
 
 class FirstPaymentBuilderTest extends BaseTestCase
@@ -45,11 +46,13 @@ class FirstPaymentBuilderTest extends BaseTestCase
             new AddBalance(
                 $owner,
                 money(500, 'EUR'),
+                1,
                 'Test add balance 1'
             ),
             new AddBalance(
                 $owner,
                 money(500, 'EUR'),
+                1,
                 'Test add balance 2'
             ),
         ]);
@@ -69,27 +72,7 @@ class FirstPaymentBuilderTest extends BaseTestCase
             'metadata' => [
                 'owner' => [
                     'type' => get_class($owner),
-                    'id' => $owner->id,
-                ],
-                'actions' => [
-                    [
-                        'handler' => AddBalance::class,
-                        'subtotal' => [
-                            'value' => '5.00',
-                            'currency' => 'EUR',
-                        ],
-                        'taxPercentage' => 0,
-                        'description' => 'Test add balance 1',
-                    ],
-                    [
-                        'handler' => AddBalance::class,
-                        'subtotal' => [
-                            'value' => '5.00',
-                            'currency' => 'EUR',
-                        ],
-                        'taxPercentage' => 0,
-                        'description' => 'Test add balance 2',
-                    ],
+                    'id' => $owner->getKey(),
                 ],
             ],
         ];
@@ -116,17 +99,33 @@ class FirstPaymentBuilderTest extends BaseTestCase
             new AddBalance(
                 $owner,
                 money(500, 'EUR'),
+                1,
                 'Test add balance 1'
             ),
             new AddBalance(
                 $owner,
                 money(500, 'EUR'),
+                1,
                 'Test add balance 2'
             ),
         ]);
 
         $this->mock(CreateMolliePayment::class, function (CreateMolliePayment $mock) {
-            $payment = new Payment(new MollieApiClient);
+            $payment = new MolliePayment(new MollieApiClient);
+            $payment->id = 'tr_unique_id';
+            $payment->amount = (object) [
+                'currency' => 'EUR',
+                'value' => '12.34',
+            ];
+            $payment->amountChargedBack = (object) [
+                'currency' => 'EUR',
+                'value' => '0.00',
+            ];
+            $payment->amountRefunded = (object) [
+                'currency' => 'EUR',
+                'value' => '0.00',
+            ];
+            $payment->mandateId = 'mdt_dummy_mandate_id';
 
             return $mock->shouldReceive('execute')
                 ->once()
@@ -138,7 +137,7 @@ class FirstPaymentBuilderTest extends BaseTestCase
         $this->assertEquals(0, $owner->orderItems()->count());
         $this->assertEquals(0, $owner->orders()->count());
 
-        $this->assertInstanceOf(Payment::class, $payment);
+        $this->assertInstanceOf(MolliePayment::class, $payment);
     }
 
     /** @test */
@@ -151,9 +150,22 @@ class FirstPaymentBuilderTest extends BaseTestCase
         ]);
 
         $this->mock(CreateMolliePayment::class, function (CreateMolliePayment $mock) {
-            $payment = new Payment(new MollieApiClient);
+            $payment = new MolliePayment(new MollieApiClient);
             $payment->redirectUrl = 'https://www.example.com/{payment_id}';
             $payment->id = 'tr_unique_id';
+            $payment->amount = (object) [
+                'currency' => 'EUR',
+                'value' => '12.34',
+            ];
+            $payment->amountRefunded = (object) [
+                'currency' => 'EUR',
+                'value' => '0.00',
+            ];
+            $payment->amountChargedBack = (object) [
+                'currency' => 'EUR',
+                'value' => '0.00',
+            ];
+            $payment->mandateId = 'mdt_dummy_mandate_id';
 
             return $mock->shouldReceive('execute')
                 ->once()
@@ -161,9 +173,63 @@ class FirstPaymentBuilderTest extends BaseTestCase
         });
 
         $payment = $builder->inOrderTo([
-            new AddGenericOrderItem($owner, money(100, 'EUR'), 'Parse redirectUrl test'),
+            new AddGenericOrderItem($owner, money(100, 'EUR'), 1, 'Parse redirectUrl test'),
         ])->create();
 
         $this->assertEquals('https://www.example.com/tr_unique_id', $payment->redirectUrl);
+    }
+
+    /** @test */
+    public function storesLocalPaymentRecord()
+    {
+        $owner = factory(User::class)->create();
+        $this->assertEquals(0, $owner->orderItems()->count());
+        $this->assertEquals(0, $owner->orders()->count());
+
+        $builder = new FirstPaymentBuilder($owner, [
+            'description' => 'Test mandate payment',
+            'redirectUrl' => 'https://www.example.com',
+        ]);
+
+        $builder->inOrderTo([
+            new AddBalance(
+                $owner,
+                money(500, 'EUR'),
+                1,
+                'Test add balance 1'
+            ),
+            new AddBalance(
+                $owner,
+                money(500, 'EUR'),
+                1,
+                'Test add balance 2'
+            ),
+        ]);
+
+        $this->mock(CreateMolliePayment::class, function (CreateMolliePayment $mock) {
+            $payment = new MolliePayment(new MollieApiClient);
+            $payment->id = 'tr_dummy_payment_id';
+            $payment->amount = (object) [
+                'currency' => 'EUR',
+                'value' => '12.34',
+            ];
+            $payment->mandateId = 'mdt_dummy_mandate_id';
+
+            return $mock->shouldReceive('execute')
+                ->once()
+                ->andReturn($payment);
+        });
+
+        $molliePayment = $builder->create();
+
+        $localPayment = Payment::findByPaymentIdOrFail($molliePayment->id);
+        $this->assertNull($localPayment->order_id);
+        $this->assertEquals('tr_dummy_payment_id', $localPayment->mollie_payment_id);
+        $this->assertEquals('open', $localPayment->mollie_payment_status);
+        $this->assertTrue($localPayment->owner->is($owner));
+        $this->assertEquals('EUR', $localPayment->currency);
+        $this->assertEquals(1234, $localPayment->amount);
+        $this->assertEquals(0, $localPayment->amount_refunded);
+        $this->assertEquals(0, $localPayment->amount_charged_back);
     }
 }
